requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
MarkupSafe==1.1.1
python-dotenv==0.14.0
requests==2.20.0
urllib3==1.24.3
Werkzeug==1.0.1
